Fix the infinity hasu keymap

+++ keyboards/infinity_chibios/keymaps/hasu/keymap.c
@@ -1,6 +1,6 @@
 #include "infinity_chibios.h"
 
-const uint8_t keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
+const uint16_t keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
     /* Layer 0: Default Layer
      * ,-----------------------------------------------------------.
      * |Esc|  1|  2|  3|  4|  5|  6|  7|  8|  9|  0|  -|  =|  `|BSp|
